MQTT Plaintext Demo Port from FreeRTOS/FreeRTOS (#2506)
* Port MQTT Plaintext demo from FreeRTOS/FreeRTOS
- Add to demo running infra (iot_demo_runner.h and config file macro).
- Add mqtt_demo_plaintext_config.h to all boards.
- Change all transport_freertos uses to transport_secure_sockets.
- Return EXIT_SUCCESS and EXIT_FAILURE, which needs functions to not assert on non-fatal errors.

* Opportunistic warnings fixes and other fixes.
- Update comments comments in https demo CMakeLists.txt.
- Fix some warnings in transport_secure_sockets and iot_tls
- Delete autogenerated Visual Studio project file.

+++ libraries/abstractions/transport/secure_sockets/transport_secure_sockets.c
@@ -116,7 +116,7 @@ int32_t SecureSocketsTransport_Send( const NetworkContext_t * pNetworkContext,
         /* If an error occurred, a negative value is returned. @ref SocketsErrors. */
         if( bytesSent >= 0 )
         {
-            if( bytesSent < bytesToSend )
+            if( ( size_t ) bytesSent < bytesToSend )
             {
                 LogWarn( ( "bytesSent %ld < bytesToSend %ld.", bytesSent, bytesToSend ) );
             }
@@ -176,7 +176,7 @@ int32_t SecureSocketsTransport_Recv( const NetworkContext_t * pNetworkContext,
         }
         else
         {
-            if( bytesReceived < bytesToRecv )
+            if( ( size_t ) bytesReceived < bytesToRecv )
             {
                 LogInfo( ( "Receive requested %lu bytes, but %lu bytes received instead.",
                            bytesToRecv,
